Order book

+++ src/main/java/za/co/valr/valrtest/repository/OrderBookRepository.java
@@ -2,6 +2,6 @@ package za.co.valr.valrtest.repository;
 import org.springframework.data.repository.CrudRepository;
 import za.co.valr.valrtest.model.OrderBookEntity;
 
-public interface OrderBookRepository extends CrudRepository<OrderBookEntity,Integer> {
+public interface OrderBookRepository extends CrudRepository<OrderBookEntity,Long> {
     public OrderBookEntity findByCurrencyPair(String currencyPair);
 }
